Add files via upload
A more detailed description of the file you are uploading. For example, if you are uploading a document, you could specify the document title, author, or keywords.
Instructions on how to use the file. For example, if you are uploading a spreadsheet, you could describe the format of the data or how to use the formulas.
Additional information about the file that is not included in the filename. For example, if you are uploading an image, you could describe the content of the image or the purpose of the image.
In essence, you can use this text box to provide any additional context or information that would be helpful to someone who downloads the file.

Here are some specific examples of what you could write in the text box for a port scanner script:

"A simple Python script for scanning ports on a target host. Enter the target host, starting port, and ending port to identify open ports."
"This script scans a specified range of ports (TCP) on a given host. Use responsibly and ethically."
"Basic port scanner written in Python. Requires Python 3.x and understanding of network concepts (ports, protocols)."